Refactor to reduce one public method
