+++ tests/Unit/Models/ZoneTest.php
@@ -149,51 +149,4 @@ class ZoneTest extends TestCase
 
         $this->assertEquals(ZoneType::Secondary, $zone->getTypeOfZone());
     }
-
-    /**
-     * @test
-     *
-     * @dataProvider providesSerialNumberTestCases
-     */
-    public function it_should_calculate_the_serial_number(
-        string $date,
-        int $input,
-        int $expected,
-    ): void {
-        /** @var Zone $zone */
-        $zone = Zone::factory()->make([
-            'serial' => $input,
-        ]);
-
-        $this->travelTo(Carbon::createFromFormat('Y-m-d', $date));
-
-        $this->assertEquals($expected, $zone->calculateNewSerialNumber());
-    }
-
-    public static function providesSerialNumberTestCases(): \Generator
-    {
-        yield 'serial number was not changed today' => [
-            'date' => '2022-01-01',
-            'input' => 2019010100,
-            'expected' => 2022010100,
-        ];
-
-        yield 'serial number was changed only once today' => [
-            'date' => '2022-01-01',
-            'input' => 2022010100,
-            'expected' => 2022010101,
-        ];
-
-        yield 'serial number was changed several times today' => [
-            'date' => '2022-01-01',
-            'input' => 2022010109,
-            'expected' => 2022010110,
-        ];
-
-        yield 'serial number was changed 98 times today' => [
-            'date' => '2022-01-01',
-            'input' => 2022010199,
-            'expected' => 2022010200,
-        ];
-    }
 }
